lipikaar_website/backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
backports.functools-lru-cache==2.0.0
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.2
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-object-actions==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==21.2.0
idna==3.6
indic-transliteration==2.3.57
kombu==5.3.5
numpy==1.23.5
opencv-contrib-python==4.5.1.48
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.72
packaging==23.2
pdf2image==1.16.2
pillow==10.2.0
pip==23.0.1
prompt-toolkit==3.0.43
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
redis==5.0.1
regex==2023.12.25
reportlab==3.6.8
requests==2.31.0
roman==4.1
setuptools==56.0.0
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
typer==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
